Moved tests to custom/ folder
